requirements.txt
altgraph==0.17.4
comtypes==1.4.13
et_xmlfile==2.0.0
joblib==1.5.2
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
pillow==12.0.0
pygrabber==0.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
